Streamlined split updates, optimization/writing to files now thread-safe
*	Added a semaphore in each bush to determine whether it is busy being
	written or equilibrated
*	only update splits once per bush when equilibrating
*	write bush to file immediately after equilibrating
*	added method to determine default bush file location
*	removed bulk bush writing
*	other minor optimizations

+++ src/edu/utexas/wrap/util/calc/TSGCCalculator.java
@@ -26,7 +26,7 @@ public class TSGCCalculator extends Thread {
 			for (AssignmentContainer b : o.getContainers()) {
 				Map<Link,Double> flows = b.getFlows();
 				for (Link l : b.getLinks()) {
-					tsgc += flows.getOrDefault(l, 0.0).doubleValue() * l.getPrice(b.getVOT(),b.getVehicleClass()).doubleValue();
+					tsgc += flows.getOrDefault(l, 0.0) * l.getPrice(b.getVOT(),b.getVehicleClass());
 				}
 			}
 		}
